--- Projects/UOContent/World Saves/ArchiveSaves.cs
@@ -1,10 +1,0 @@
-using System;
-
-namespace Server.Saves
-{
-    public static class ArchiveSaves
-    {
-        public static Action<DateTime> Archive { get; set; }
-        public static Action<DateTime> Prune { get; set; }
-    }
-}
